@@ -26,20 +26,20 @@
   "keywords": [],
   "author": "",
   "dependencies": {
-    "@angular/animations": "^17.3.6",
-    "@angular/cdk": "^17.3.6",
-    "@angular/cli": "^17.3.6",
-    "@angular/common": "^17.3.6",
-    "@angular/compiler": "^17.3.6",
-    "@angular/core": "^17.3.6",
-    "@angular/forms": "^17.3.6",
-    "@angular/material": "^17.3.6",
-    "@angular/material-moment-adapter": "^17.3.6",
-    "@angular/platform-browser": "^17.3.6",
-    "@angular/platform-browser-dynamic": "^17.3.6",
-    "@angular/router": "^17.3.6",
-    "@angular/service-worker": "^17.3.6",
-    "@angular/upgrade": "^17.3.6",
+    "@angular/animations": "^18.0.3",
+    "@angular/cdk": "^18.0.1",
+    "@angular/cli": "^18.0.4",
+    "@angular/common": "^18.0.3",
+    "@angular/compiler": "^18.0.3",
+    "@angular/core": "^18.0.3",
+    "@angular/forms": "^18.0.3",
+    "@angular/material": "^18.0.1",
+    "@angular/material-moment-adapter": "^18.0.1",
+    "@angular/platform-browser": "^18.0.3",
+    "@angular/platform-browser-dynamic": "^18.0.3",
+    "@angular/router": "^18.0.3",
+    "@angular/service-worker": "^18.0.3",
+    "@angular/upgrade": "^18.0.3",
     "@ctrl/ngx-emoji-mart": "^9.2.0",
     "@ngneat/hotkeys": "^4.0.0",
     "@uirouter/angular": "^13.0",
@@ -154,11 +151,10 @@
     "postcss": "^8.4.27",
     "postcss-scss": "^0.1.7",
     "prettier": "^3.1.0",
-    "protractor": "~7.0.0",
     "sass": "^1.48.0",
     "tailwindcss": "~3.3",
     "ts-node": "~10.9",
-    "typescript": "~5.2",
+    "typescript": "~5.4.5",
     "underscore": "^1.8.3"
   },
   "optionalDependencies": {
