@@ -5,7 +5,7 @@ import {NgModule, Injector, DoBootstrap} from '@angular/core';
 import {BrowserModule, DomSanitizer, Title} from '@angular/platform-browser';
 import {UpgradeModule} from '@angular/upgrade/static';
 import {AppInjector, setAppInjector} from './app-injector';
-import {HttpClientModule, HTTP_INTERCEPTORS} from '@angular/common/http';
+import { HTTP_INTERCEPTORS, provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 import {BrowserAnimationsModule} from '@angular/platform-browser/animations';
 
 // Lottie animation module
@@ -225,243 +225,237 @@ import {TasksViewerComponent} from './units/states/tasks/tasks-viewer/tasks-view
 import {UnitCodeComponent} from './common/unit-code/unit-code.component';
 import {GradeService} from './common/services/grade.service';
 
-@NgModule({
-  // Components we declare
-  declarations: [
-    AlertComponent,
-    AboutDoubtfireModalContent,
-    TeachingPeriodUnitImportDialogComponent,
-    TaskCommentComposerComponent,
-    AudioCommentRecorderComponent,
-    MicrophoneTesterComponent,
-    DiscussionPromptComposerComponent,
-    IntelligentDiscussionPlayerComponent,
-    IntelligentDiscussionDialog,
-    DiscussionComposerDialog,
-    IntelligentDiscussionRecorderComponent,
-    ExtensionCommentComponent,
-    PdfImageCommentComponent,
-    CampusListComponent,
-    ActivityTypeListComponent,
-    OverseerImageListComponent,
-    ExtensionModalComponent,
-    CalendarModalComponent,
-    InstitutionSettingsComponent,
-    HomeComponent,
-    CommentBubbleActionComponent,
-    UnitTutorialsListComponent,
-    UnitTutorialsManagerComponent,
-    FileDropComponent,
-    UnitStudentsEditorComponent,
-    UnitTaskEditorComponent,
-    TaskDefinitionEditorComponent,
-    TaskDefinitionGeneralComponent,
-    TaskDefinitionWhoComponent,
-    TaskDefinitionDatesComponent,
-    TaskDefinitionUploadComponent,
-    TaskDefinitionOptionsComponent,
-    TaskDefinitionResourcesComponent,
-    TaskDefinitionOverseerComponent,
-    UnitAnalyticsComponent,
-    StudentTutorialSelectComponent,
-    StudentCampusSelectComponent,
-    TaskListItemComponent,
-    CreatePortfolioTaskListItemComponent,
-    TaskDescriptionCardComponent,
-    StatusIconComponent,
-    TaskCommentsViewerComponent,
-    UserIconComponent,
-    AudioPlayerComponent,
-    MarkedPipe,
-    HumanizedDatePipe,
-    IsActiveUnitRole,
-    DragDropDirective,
-    fPdfViewerComponent,
-    SafePipe,
-    PdfViewerPanelComponent,
-    StaffTaskListComponent,
-    TaskSimilarityViewComponent,
-    FiltersPipe,
-    TasksOfTaskDefinitionPipe,
-    TasksInTutorialsPipe,
-    TasksForInboxSearchPipe,
-    StatusIconComponent,
-    TaskAssessmentCommentComponent,
-    TaskAssessmentModalComponent,
-    TaskSubmissionHistoryComponent,
-    HeaderComponent,
-    UnitDropdownComponent,
-    TaskDropdownComponent,
-    SplashScreenComponent,
-    ObjectSelectComponent,
-    WelcomeComponent,
-    AcceptEulaComponent,
-    HeroSidebarComponent,
-    SignInComponent,
-    EditProfileFormComponent,
-    EditProfileComponent,
-    UserBadgeComponent,
-    TaskStatusCardComponent,
-    TaskDueCardComponent,
-    FooterComponent,
-    TaskAssessmentCardComponent,
-    TaskSubmissionCardComponent,
-    TaskDashboardComponent,
-    InboxComponent,
-    ProjectProgressBarComponent,
-    TeachingPeriodListComponent,
-    CreateNewUnitModal,
-    CreateNewUnitModalContentComponent,
-    TiiActionLogComponent,
-    FChipComponent,
-    UnitCodeComponent,
-    NewTeachingPeriodDialogComponent,
-    FileViewerComponent,
-    AlertComponent,
-    FUnitTaskListComponent,
-    FTaskDetailsViewComponent,
-    FTaskSheetViewComponent,
-    TasksViewerComponent,
-    FUsersComponent,
-    FTaskBadgeComponent,
-    FUnitsComponent,
-  ],
-  // Services we provide
-  providers: [
-    AlertService,
-    MarkedPipe,
-    CampusService,
-    AuthenticationService,
-    GroupSetService,
-    GroupService,
-    UnitService,
-    ProjectService,
-    UnitRoleService,
-    LearningOutcomeService,
-    TaskDefinitionService,
-    TeachingPeriodService,
-    TiiActionService,
-    TeachingPeriodBreakService,
-    TeachingPeriodUnitImportService,
-    TutorialService,
-    TutorialStreamService,
-    UserService,
-    TaskService,
-    GradeService,
-    TaskSimilarityService,
-    WebcalService,
-    ActivityTypeService,
-    OverseerImageService,
-    OverseerAssessmentService,
-    EmojiService,
-    FileDownloaderService,
-    CheckForUpdateService,
-    TaskOutcomeAlignmentService,
-    visualisationsProvider,
-    commentsModalProvider,
-    rootScopeProvider,
-    calendarModalProvider,
-    aboutDoubtfireModalProvider,
-    uploadSubmissionModalProvider,
-    gradeTaskModalProvider,
-    analyticsServiceProvider,
-    dateServiceProvider,
-    CsvUploadModalProvider,
-    CsvResultModalProvider,
-    {provide: MAT_DATE_LOCALE, useValue: 'en-AU'},
-    UnitStudentEnrolmentModalProvider,
-    TaskCommentService,
-    AudioRecorderProvider,
-    AudioRecorderServiceProvider,
-    plagiarismReportModalProvider,
-    UnitStudentsEditorComponent,
-    ConfirmationModalProvider,
-    {
-      provide: HTTP_INTERCEPTORS,
-      useClass: HttpAuthenticationInterceptor,
-      multi: true,
-      deps: [UserService],
-    },
-    {
-      provide: HTTP_INTERCEPTORS,
-      useClass: HttpErrorInterceptor,
-      multi: true,
-      deps: [AuthenticationService, UserService],
-    },
-    AboutDoubtfireModal,
-    AboutDoubtfireModalService,
-    DoubtfireConstants,
-    TasksOfTaskDefinitionPipe,
-    TasksInTutorialsPipe,
-    TasksForInboxSearchPipe,
-    IsActiveUnitRole,
-    CreateNewUnitModal,
-    provideLottieOptions({
-      player: () => player,
-    }),
-  ],
-  imports: [
-    FlexLayoutModule,
-    BrowserModule,
-    BrowserAnimationsModule,
-    FormsModule,
-    HttpClientModule,
-    ClipboardModule,
-    DragDropModule,
-    ScrollingModule,
-    MatToolbarModule,
-    MatFormFieldModule,
-    MatAutocompleteModule,
-    MatInputModule,
-    MatBadgeModule,
-    MatRadioModule,
-    MatListModule,
-    MatOptionModule,
-    MatStepperModule,
-    MatPaginatorModule,
-    MatSelectModule,
-    MatNativeDateModule,
-    MatButtonToggleModule,
-    MatTooltipModule,
-    MatSlideToggleModule,
-    MatButtonModule,
-    MatMenuModule,
-    MatCardModule,
-    MatCheckboxModule,
-    MatDividerModule,
-    MatDialogModule,
-    MatSortModule,
-    MatProgressBarModule,
-    MatIconModule,
-    MatProgressSpinnerModule,
-    MatSliderModule,
-    MatDatepickerModule,
-    MatExpansionModule,
-    MatCardModule,
-    MatGridListModule,
-    MatSelectModule,
-    MatToolbarModule,
-    MatTabsModule,
-    UpgradeModule,
-    MatTableModule,
-    MatTabsModule,
-    MatChipsModule,
-    MatSnackBarModule,
-    ReactiveFormsModule,
-    PickerModule,
-    EmojiModule,
-    PdfViewerModule,
-    LottieComponent,
-    UIRouterUpgradeModule.forRoot({states: doubtfireStates}),
-    ServiceWorkerModule.register('ngsw-worker.js', {
-      enabled: environment.production,
-      registrationStrategy: () => interval(6000).pipe(take(1)),
-    }),
-    MatDatepickerModule,
-    MatNativeDateModule,
-    MatDialogModuleNew,
-  ],
-})
+@NgModule({ 
+    // Components we declare
+    declarations: [
+        AlertComponent,
+        AboutDoubtfireModalContent,
+        TeachingPeriodUnitImportDialogComponent,
+        TaskCommentComposerComponent,
+        AudioCommentRecorderComponent,
+        MicrophoneTesterComponent,
+        DiscussionPromptComposerComponent,
+        IntelligentDiscussionPlayerComponent,
+        IntelligentDiscussionDialog,
+        DiscussionComposerDialog,
+        IntelligentDiscussionRecorderComponent,
+        ExtensionCommentComponent,
+        PdfImageCommentComponent,
+        CampusListComponent,
+        ActivityTypeListComponent,
+        OverseerImageListComponent,
+        ExtensionModalComponent,
+        CalendarModalComponent,
+        InstitutionSettingsComponent,
+        HomeComponent,
+        CommentBubbleActionComponent,
+        UnitTutorialsListComponent,
+        UnitTutorialsManagerComponent,
+        FileDropComponent,
+        UnitStudentsEditorComponent,
+        UnitTaskEditorComponent,
+        TaskDefinitionEditorComponent,
+        TaskDefinitionGeneralComponent,
+        TaskDefinitionWhoComponent,
+        TaskDefinitionDatesComponent,
+        TaskDefinitionUploadComponent,
+        TaskDefinitionOptionsComponent,
+        TaskDefinitionResourcesComponent,
+        TaskDefinitionOverseerComponent,
+        UnitAnalyticsComponent,
+        StudentTutorialSelectComponent,
+        StudentCampusSelectComponent,
+        TaskListItemComponent,
+        CreatePortfolioTaskListItemComponent,
+        TaskDescriptionCardComponent,
+        StatusIconComponent,
+        TaskCommentsViewerComponent,
+        UserIconComponent,
+        AudioPlayerComponent,
+        MarkedPipe,
+        HumanizedDatePipe,
+        IsActiveUnitRole,
+        DragDropDirective,
+        fPdfViewerComponent,
+        SafePipe,
+        PdfViewerPanelComponent,
+        StaffTaskListComponent,
+        TaskSimilarityViewComponent,
+        FiltersPipe,
+        TasksOfTaskDefinitionPipe,
+        TasksInTutorialsPipe,
+        TasksForInboxSearchPipe,
+        StatusIconComponent,
+        TaskAssessmentCommentComponent,
+        TaskAssessmentModalComponent,
+        TaskSubmissionHistoryComponent,
+        HeaderComponent,
+        UnitDropdownComponent,
+        TaskDropdownComponent,
+        SplashScreenComponent,
+        ObjectSelectComponent,
+        WelcomeComponent,
+        AcceptEulaComponent,
+        HeroSidebarComponent,
+        SignInComponent,
+        EditProfileFormComponent,
+        EditProfileComponent,
+        UserBadgeComponent,
+        TaskStatusCardComponent,
+        TaskDueCardComponent,
+        FooterComponent,
+        TaskAssessmentCardComponent,
+        TaskSubmissionCardComponent,
+        TaskDashboardComponent,
+        InboxComponent,
+        ProjectProgressBarComponent,
+        TeachingPeriodListComponent,
+        CreateNewUnitModal,
+        CreateNewUnitModalContentComponent,
+        TiiActionLogComponent,
+        FChipComponent,
+        UnitCodeComponent,
+        NewTeachingPeriodDialogComponent,
+        FileViewerComponent,
+        AlertComponent,
+        FUnitTaskListComponent,
+        FTaskDetailsViewComponent,
+        FTaskSheetViewComponent,
+        TasksViewerComponent,
+        FUsersComponent,
+        FTaskBadgeComponent,
+        FUnitsComponent,
+    ], imports: [FlexLayoutModule,
+        BrowserModule,
+        BrowserAnimationsModule,
+        FormsModule,
+        ClipboardModule,
+        DragDropModule,
+        ScrollingModule,
+        MatToolbarModule,
+        MatFormFieldModule,
+        MatAutocompleteModule,
+        MatInputModule,
+        MatBadgeModule,
+        MatRadioModule,
+        MatListModule,
+        MatOptionModule,
+        MatStepperModule,
+        MatPaginatorModule,
+        MatSelectModule,
+        MatNativeDateModule,
+        MatButtonToggleModule,
+        MatTooltipModule,
+        MatSlideToggleModule,
+        MatButtonModule,
+        MatMenuModule,
+        MatCardModule,
+        MatCheckboxModule,
+        MatDividerModule,
+        MatDialogModule,
+        MatSortModule,
+        MatProgressBarModule,
+        MatIconModule,
+        MatProgressSpinnerModule,
+        MatSliderModule,
+        MatDatepickerModule,
+        MatExpansionModule,
+        MatCardModule,
+        MatGridListModule,
+        MatSelectModule,
+        MatToolbarModule,
+        MatTabsModule,
+        UpgradeModule,
+        MatTableModule,
+        MatTabsModule,
+        MatChipsModule,
+        MatSnackBarModule,
+        ReactiveFormsModule,
+        PickerModule,
+        EmojiModule,
+        PdfViewerModule,
+        LottieComponent,
+        UIRouterUpgradeModule.forRoot({ states: doubtfireStates }),
+        ServiceWorkerModule.register('ngsw-worker.js', {
+            enabled: environment.production,
+            registrationStrategy: () => interval(6000).pipe(take(1)),
+        }),
+        MatDatepickerModule,
+        MatNativeDateModule,
+        MatDialogModuleNew], providers: [
+        AlertService,
+        MarkedPipe,
+        CampusService,
+        AuthenticationService,
+        GroupSetService,
+        GroupService,
+        UnitService,
+        ProjectService,
+        UnitRoleService,
+        LearningOutcomeService,
+        TaskDefinitionService,
+        TeachingPeriodService,
+        TiiActionService,
+        TeachingPeriodBreakService,
+        TeachingPeriodUnitImportService,
+        TutorialService,
+        TutorialStreamService,
+        UserService,
+        TaskService,
+        GradeService,
+        TaskSimilarityService,
+        WebcalService,
+        ActivityTypeService,
+        OverseerImageService,
+        OverseerAssessmentService,
+        EmojiService,
+        FileDownloaderService,
+        CheckForUpdateService,
+        TaskOutcomeAlignmentService,
+        visualisationsProvider,
+        commentsModalProvider,
+        rootScopeProvider,
+        calendarModalProvider,
+        aboutDoubtfireModalProvider,
+        uploadSubmissionModalProvider,
+        gradeTaskModalProvider,
+        analyticsServiceProvider,
+        dateServiceProvider,
+        CsvUploadModalProvider,
+        CsvResultModalProvider,
+        { provide: MAT_DATE_LOCALE, useValue: 'en-AU' },
+        UnitStudentEnrolmentModalProvider,
+        TaskCommentService,
+        AudioRecorderProvider,
+        AudioRecorderServiceProvider,
+        plagiarismReportModalProvider,
+        UnitStudentsEditorComponent,
+        ConfirmationModalProvider,
+        {
+            provide: HTTP_INTERCEPTORS,
+            useClass: HttpAuthenticationInterceptor,
+            multi: true,
+            deps: [UserService],
+        },
+        {
+            provide: HTTP_INTERCEPTORS,
+            useClass: HttpErrorInterceptor,
+            multi: true,
+            deps: [AuthenticationService, UserService],
+        },
+        AboutDoubtfireModal,
+        AboutDoubtfireModalService,
+        DoubtfireConstants,
+        TasksOfTaskDefinitionPipe,
+        TasksInTutorialsPipe,
+        TasksForInboxSearchPipe,
+        IsActiveUnitRole,
+        CreateNewUnitModal,
+        provideLottieOptions({
+            player: () => player,
+        }),
+        provideHttpClient(withInterceptorsFromDi()),
+    ] })
 
 // There is no longer any requirement for an EntryComponents section
 // since Angular 9 introduced the IVY renderer
